+++ changelog.md
@@ -3,6 +3,8 @@
 - Cleanup logger
 - Move bind_collector to platforms.entity
 - Move config/value_property to platforms.decorators
+- Move measure_execution_time to decorators
+- Update requirements
 
 # Version 2023.9.1 (2023-09-06)